packages/sample-package/pyproject.toml
[tool.poetry]
name = "sample-package"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "sample_package"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.7"

[tool.poetry.group.test.dependencies]
test_utils = { path = "../libs/test-utils", develop = true }
anyio = "^4.0.0"
pytest-asyncio = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
